service for all the points
if you pass no params, you get everything

turn that on from the lower level library

Thanks to me from years ago for adding that

The test is sucky, and suddenly I realize the library is fragile because
I've hardcoded in the db names and tables and such.  Need to migrate
this to the new config.json way I'm doing this, which allows for things
to change over time.

Also, the test should load up known data into a test db, not test
against the live db